refactor: identify sql as specific to metadata table
